requirements.txt
giturlparse==0.10.0
PyGithub==1.55
GitPython==3.1.27
